+++ CHANGELOG.rst
@@ -5,6 +5,23 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Calendar Versioning <https://calver.org/>`_ with the schema MAJOR.MINOR.YYYY0M0D.
 
 
+0.3.20240627 - 2024-06-27
+-------------------------
+
+Added
+^^^^^
+- WarmCookie Backdoor (aka Quickbind / BadSpace)
+
+Changed
+^^^^^^^
+- Upgrade and pin dependencies:
+    - `mwcp` to `3.14.0`
+    - `dragodis` to `1.0.0`
+    - `rugosa` to `1.0.0`
+    - `pyhidra` to `1.2.0`
+- Loosen `construct` dependency to `>=2.9.45`
+
+
 0.2.20230803 - 2023-08-03
 -------------------------